upgrade to iroh 0.8.0 & move `Doc` & `Entry` into `doc.rs`
