+++ Cargo.toml
@@ -30,7 +30,7 @@ anyhow = "1.0.69"
 blake3 = "1.3.3"
 bytes = "1"
 data-encoding = { version = "2.3.3" }
-iroh = { version = "0.7.0", default-features = false, features = ["mem-db", "flat-db", "metrics"] }
+iroh = { version = "0.8.0", default-features = false, features = ["flat-db", "metrics"] }
 iroh-io = { version = "0.2.1" }
 libc = "0.2.141"
 multibase = { version = "0.9.1"}
