+++ src/node.rs
@@ -1,55 +1,61 @@
-use std::{collections::HashMap, path::PathBuf, sync::Arc, time::SystemTime};
-
-use futures::{
-    stream::{StreamExt, TryStreamExt},
-    Future,
+use std::{
+    collections::HashMap,
+    path::PathBuf,
+    sync::Arc,
+    time::{Duration, SystemTime},
 };
+
+use futures::stream::TryStreamExt;
 use iroh::{
-    baomap::flat,
     bytes::util::runtime::Handle,
-    client::Doc as ClientDoc,
     net::key::SecretKey,
     node::{Node, DEFAULT_BIND_ADDR},
-    rpc_protocol::{ProviderRequest, ProviderResponse, ShareMode},
+    rpc_protocol::{ProviderRequest, ProviderResponse},
 };
 use quic_rpc::transport::flume::FlumeConnection;
 
+use crate::block_on;
+use crate::doc::{AuthorId, Doc, DocTicket, Entry, NamespaceId};
 use crate::error::IrohError as Error;
 use crate::key::PublicKey;
-use crate::net::SocketAddr;
 
 pub use iroh::rpc_protocol::CounterStats;
-pub use iroh::sync_engine::LiveStatus;
+
+/// Information about a direct address.
+/// TODO: when refactoring the iroh.node API, give this an impl
+#[derive(Debug)]
+pub struct DirectAddrInfo(iroh::net::magicsock::DirectAddrInfo);
 
 #[derive(Debug)]
 pub struct ConnectionInfo {
-    pub id: u64,
+    /// The public key of the endpoint.
     pub public_key: Arc<PublicKey>,
+    /// Derp region, if available.
     pub derp_region: Option<u16>,
-    pub addrs: Vec<Arc<SocketAddr>>,
-    pub latencies: Vec<Option<f64>>,
+    /// List of addresses at which this node might be reachable, plus any latency information we
+    /// have about that address and the last time the address was used.
+    pub addrs: Vec<Arc<DirectAddrInfo>>,
+    /// The type of connection we have to the peer, either direct or over relay.
     pub conn_type: ConnectionType,
-    pub latency: Option<f64>,
+    /// The latency of the `conn_type`.
+    pub latency: Option<Duration>,
+    /// Duration since the last time this peer was used.
+    pub last_used: Option<Duration>,
 }
 
 impl From<iroh::net::magic_endpoint::ConnectionInfo> for ConnectionInfo {
     fn from(value: iroh::net::magic_endpoint::ConnectionInfo) -> Self {
         ConnectionInfo {
-            id: value.id as _,
             public_key: Arc::new(PublicKey(value.public_key)),
             derp_region: value.derp_region,
             addrs: value
                 .addrs
                 .iter()
-                .map(|(a, _)| Arc::new((*a).into()))
-                .collect(),
-            latencies: value
-                .addrs
-                .iter()
-                .map(|(_, l)| l.map(|d| d.as_secs_f64()))
+                .map(|a| Arc::new(DirectAddrInfo(a.clone())))
                 .collect(),
             conn_type: value.conn_type.into(),
-            latency: value.latency.map(|l| l.as_secs_f64()),
+            latency: value.latency,
+            last_used: value.last_used,
         }
     }
 }
@@ -58,6 +64,7 @@ impl From<iroh::net::magic_endpoint::ConnectionInfo> for ConnectionInfo {
 pub enum ConnectionType {
     Direct { addr: String, port: u16 },
     Relay { port: u16 },
+    Mixed { addr: String, port: u16 },
     None,
 }
 
@@ -67,6 +74,10 @@ impl From<iroh::net::magicsock::ConnectionType> for ConnectionType {
             iroh::net::magicsock::ConnectionType::Direct(addr) => ConnectionType::Direct {
                 addr: addr.ip().to_string(),
                 port: addr.port(),
+            },
+            iroh::net::magicsock::ConnectionType::Mixed(addr, port) => ConnectionType::Mixed {
+                addr: addr.ip().to_string(),
+                port: port,
             },
             iroh::net::magicsock::ConnectionType::Relay(port) => ConnectionType::Relay { port },
             iroh::net::magicsock::ConnectionType::None => ConnectionType::None,
@@ -102,8 +113,6 @@ pub enum LiveEvent {
     NeighborDown(PublicKey),
     /// A set-reconciliation sync finished.
     SyncFinished(SyncEvent),
-    /// The document was closed. No further events will be emitted.
-    Closed,
 }
 
 pub enum LiveEventType {
@@ -113,7 +122,6 @@ pub enum LiveEventType {
     NeighborUp,
     NeighborDown,
     SyncFinished,
-    Closed,
 }
 
 impl LiveEvent {
@@ -125,7 +133,6 @@ impl LiveEvent {
             Self::NeighborUp(_) => LiveEventType::NeighborUp,
             Self::NeighborDown(_) => LiveEventType::NeighborDown,
             Self::SyncFinished(_) => LiveEventType::SyncFinished,
-            Self::Closed => LiveEventType::Closed,
         }
     }
 
@@ -198,12 +205,12 @@ pub enum ContentStatus {
     Missing,
 }
 
-impl From<iroh::sync_engine::ContentStatus> for ContentStatus {
-    fn from(value: iroh::sync_engine::ContentStatus) -> Self {
+impl From<iroh::sync::sync::ContentStatus> for ContentStatus {
+    fn from(value: iroh::sync::sync::ContentStatus) -> Self {
         match value {
-            iroh::sync_engine::ContentStatus::Complete => Self::Complete,
-            iroh::sync_engine::ContentStatus::Incomplete => Self::Incomplete,
-            iroh::sync_engine::ContentStatus::Missing => Self::Missing,
+            iroh::sync::sync::ContentStatus::Complete => Self::Complete,
+            iroh::sync::sync::ContentStatus::Incomplete => Self::Incomplete,
+            iroh::sync::sync::ContentStatus::Missing => Self::Missing,
         }
     }
 }
@@ -229,7 +236,6 @@ impl From<iroh::sync_engine::LiveEvent> for LiveEvent {
             iroh::sync_engine::LiveEvent::NeighborUp(key) => LiveEvent::NeighborUp(key.into()),
             iroh::sync_engine::LiveEvent::NeighborDown(key) => LiveEvent::NeighborDown(key.into()),
             iroh::sync_engine::LiveEvent::SyncFinished(e) => LiveEvent::SyncFinished(e.into()),
-            iroh::sync_engine::LiveEvent::Closed => LiveEvent::Closed,
         }
     }
 }
@@ -237,14 +243,14 @@ impl From<iroh::sync_engine::LiveEvent> for LiveEvent {
 /// Outcome of a sync operation
 #[derive(Debug, Clone)]
 pub struct SyncEvent {
-    /// Namespace that was synced
-    pub namespace: Arc<NamespaceId>,
     /// Peer we synced with
     pub peer: Arc<PublicKey>,
     /// Origin of the sync exchange
     pub origin: Origin,
-    /// Timestamp when the sync finished (since unix epoch in seconds)
-    pub finished: f64,
+    /// Timestamp when the sync finished
+    pub finished: SystemTime,
+    /// Timestamp when the sync started
+    pub started: SystemTime,
     /// Result of the sync operation. `None` if successfull.
     pub result: Option<String>,
 }
@@ -252,14 +258,10 @@ pub struct SyncEvent {
 impl From<iroh::sync_engine::SyncEvent> for SyncEvent {
     fn from(value: iroh::sync_engine::SyncEvent) -> Self {
         SyncEvent {
-            namespace: Arc::new(value.namespace.into()),
             peer: Arc::new(value.peer.into()),
             origin: value.origin.into(),
-            finished: value
-                .finished
-                .duration_since(SystemTime::UNIX_EPOCH)
-                .map(|v| v.as_secs_f64())
-                .unwrap_or_default(),
+            finished: value.finished,
+            started: value.started,
             result: match value.result {
                 Ok(_) => None,
                 Err(err) => Some(err),
@@ -268,30 +270,34 @@ impl From<iroh::sync_engine::SyncEvent> for SyncEvent {
     }
 }
 
+// TODO: iroh 0.8.0 release made this struct private. Re-implement when it's made public again
 /// Why we started a sync request
-#[derive(Debug, Clone, Copy)]
-pub enum SyncReason {
-    /// Direct join request via API
-    DirectJoin,
-    /// Peer showed up as new neighbor in the gossip swarm
-    NewNeighbor,
-}
+// #[derive(Debug, Clone, Copy)]
+// pub enum SyncReason {
+//     /// Direct join request via API
+//     DirectJoin,
+//     /// Peer showed up as new neighbor in the gossip swarm
+//     NewNeighbor,
+// }
 
-impl From<iroh::sync_engine::SyncReason> for SyncReason {
-    fn from(value: iroh::sync_engine::SyncReason) -> Self {
-        match value {
-            iroh::sync_engine::SyncReason::DirectJoin => Self::DirectJoin,
-            iroh::sync_engine::SyncReason::NewNeighbor => Self::NewNeighbor,
-        }
-    }
-}
+// impl From<iroh::sync_engine::SyncReason> for SyncReason {
+//     fn from(value: iroh::sync_engine::SyncReason) -> Self {
+//         match value {
+//             iroh::sync_engine::SyncReason::DirectJoin => Self::DirectJoin,
+//             iroh::sync_engine::SyncReason::NewNeighbor => Self::NewNeighbor,
+//         }
+//     }
+// }
 
 /// Why we performed a sync exchange
 #[derive(Debug, Clone)]
 pub enum Origin {
-    Connect {
-        reason: SyncReason,
-    },
+    /// TODO: in iroh 0.8.0 `SyncReason` is private, until the next release when it can be made
+    /// public, use a unit variant
+    // Connect {
+    //     reason: SyncReason,
+    // },
+    Connect,
     /// A peer connected to us and we accepted the exchange
     Accept,
 }
@@ -299,9 +305,7 @@ pub enum Origin {
 impl From<iroh::sync_engine::Origin> for Origin {
     fn from(value: iroh::sync_engine::Origin) -> Self {
         match value {
-            iroh::sync_engine::Origin::Connect(reason) => Self::Connect {
-                reason: reason.into(),
-            },
+            iroh::sync_engine::Origin::Connect(_) => Self::Connect,
             iroh::sync_engine::Origin::Accept => Self::Accept,
         }
     }
@@ -318,7 +322,7 @@ pub struct InsertRemoteEvent {
 }
 
 #[derive(Debug, Clone)]
-pub struct Hash(iroh::bytes::Hash);
+pub struct Hash(pub(crate) iroh::bytes::Hash);
 
 impl From<iroh::bytes::Hash> for Hash {
     fn from(h: iroh::bytes::Hash) -> Self {
@@ -341,191 +345,12 @@ impl From<Hash> for iroh::bytes::Hash {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct Entry(iroh::sync::sync::Entry);
-
-impl From<iroh::sync::sync::Entry> for Entry {
-    fn from(e: iroh::sync::sync::Entry) -> Self {
-        Entry(e)
-    }
-}
-
-impl Entry {
-    pub fn author(&self) -> Arc<AuthorId> {
-        Arc::new(AuthorId(self.0.id().author()))
-    }
-
-    pub fn key(&self) -> Vec<u8> {
-        self.0.id().key().to_vec()
-    }
-
-    pub fn hash(&self) -> Arc<Hash> {
-        Arc::new(Hash(self.0.content_hash()))
-    }
-
-    pub fn namespace(&self) -> Arc<NamespaceId> {
-        Arc::new(NamespaceId(self.0.id().namespace()))
-    }
-}
-
-pub struct Doc {
-    inner: ClientDoc<FlumeConnection<ProviderResponse, ProviderRequest>>,
-    rt: Handle,
-}
-
-impl Doc {
-    pub fn id(&self) -> String {
-        self.inner.id().to_string()
-    }
-
-    pub fn keys(&self) -> Result<Vec<Arc<Entry>>, Error> {
-        let latest = block_on(&self.rt, async {
-            let get_result = self
-                .inner
-                .get_many(iroh::sync::store::GetFilter::All)
-                .await?;
-            get_result
-                .map_ok(|e| Arc::new(Entry(e)))
-                .try_collect::<Vec<_>>()
-                .await
-        })
-        .map_err(Error::doc)?;
-        Ok(latest)
-    }
-
-    pub fn share_write(&self) -> Result<Arc<DocTicket>, Error> {
-        block_on(&self.rt, async {
-            let ticket = self
-                .inner
-                .share(ShareMode::Write)
-                .await
-                .map_err(Error::doc)?;
-
-            Ok(Arc::new(DocTicket(ticket)))
-        })
-    }
-
-    pub fn share_read(&self) -> Result<Arc<DocTicket>, Error> {
-        block_on(&self.rt, async {
-            let ticket = self
-                .inner
-                .share(ShareMode::Read)
-                .await
-                .map_err(Error::doc)?;
-
-            Ok(Arc::new(DocTicket(ticket)))
-        })
-    }
-
-    pub fn set_bytes(
-        &self,
-        author_id: Arc<AuthorId>,
-        key: Vec<u8>,
-        value: Vec<u8>,
-    ) -> Result<Arc<Hash>, Error> {
-        block_on(&self.rt, async {
-            let hash = self
-                .inner
-                .set_bytes(author_id.0.clone(), key, value)
-                .await
-                .map_err(Error::doc)?;
-            Ok(Arc::new(Hash(hash)))
-        })
-    }
-
-    pub fn get_content_bytes(&self, entry: Arc<Entry>) -> Result<Vec<u8>, Error> {
-        block_on(&self.rt, async {
-            let content = self
-                .inner
-                .read_to_bytes(&entry.0)
-                .await
-                .map_err(Error::doc)?;
-
-            Ok(content.to_vec())
-        })
-    }
-
-    pub fn stop_sync(&self) -> Result<(), Error> {
-        block_on(&self.rt, async {
-            self.inner.leave().await.map_err(Error::doc)?;
-            Ok(())
-        })
-    }
-
-    pub fn status(&self) -> Result<LiveStatus, Error> {
-        block_on(&self.rt, async {
-            let status = self.inner.status().await.map_err(Error::doc)?;
-            Ok(status)
-        })
-    }
-
-    pub fn subscribe(&self, cb: Box<dyn SubscribeCallback>) -> Result<(), Error> {
-        let client = self.inner.clone();
-        self.rt.main().spawn(async move {
-            let mut sub = client.subscribe().await.unwrap();
-            while let Some(event) = sub.next().await {
-                match event {
-                    Ok(event) => {
-                        if let Err(err) = cb.event(Arc::new(event.into())) {
-                            println!("cb error: {:?}", err);
-                        }
-                    }
-                    Err(err) => {
-                        println!("rpc error: {:?}", err);
-                    }
-                }
-            }
-        });
-
-        Ok(())
-    }
-}
-
 pub trait SubscribeCallback: Send + Sync + 'static {
     fn event(&self, event: Arc<LiveEvent>) -> Result<(), Error>;
 }
 
-pub struct AuthorId(iroh::sync::sync::AuthorId);
-
-impl AuthorId {
-    pub fn to_string(&self) -> String {
-        self.0.to_string()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct NamespaceId(iroh::sync::sync::NamespaceId);
-
-impl From<iroh::sync::sync::NamespaceId> for NamespaceId {
-    fn from(id: iroh::sync::sync::NamespaceId) -> Self {
-        NamespaceId(id)
-    }
-}
-
-impl NamespaceId {
-    pub fn to_string(&self) -> String {
-        self.0.to_string()
-    }
-}
-
-#[derive(Debug)]
-pub struct DocTicket(iroh::rpc_protocol::DocTicket);
-
-impl DocTicket {
-    pub fn from_string(content: String) -> Result<Self, Error> {
-        let ticket = content
-            .parse::<iroh::rpc_protocol::DocTicket>()
-            .map_err(Error::doc_ticket)?;
-        Ok(DocTicket(ticket))
-    }
-
-    pub fn to_string(&self) -> String {
-        self.0.to_string()
-    }
-}
-
 pub struct IrohNode {
-    node: Node<flat::Store, iroh::sync::store::fs::Store>,
+    node: Node<iroh::bytes::store::flat::Store>,
     async_runtime: Handle,
     sync_client: iroh::client::Iroh<FlumeConnection<ProviderResponse, ProviderRequest>>,
     #[allow(dead_code)]
@@ -556,8 +381,10 @@ impl IrohNode {
             // create a bao store for the iroh-bytes blobs
             let blob_path = path.join("blobs");
             tokio::fs::create_dir_all(&blob_path).await?;
-            let db = iroh::baomap::flat::Store::load(&blob_path, &blob_path, &blob_path, &rt_inner)
-                .await?;
+            let db = iroh::bytes::store::flat::Store::load(
+                &blob_path, &blob_path, &blob_path, &rt_inner,
+            )
+            .await?;
 
             Node::builder(db, docs)
                 .bind_addr(DEFAULT_BIND_ADDR.into())
@@ -720,13 +547,6 @@ impl IrohNode {
             Ok(data.into())
         })
     }
-}
-
-fn block_on<F: Future<Output = T>, T>(rt: &Handle, fut: F) -> T {
-    tokio::task::block_in_place(move || match tokio::runtime::Handle::try_current() {
-        Ok(handle) => handle.block_on(fut),
-        Err(_) => rt.main().block_on(fut),
-    })
 }
 
 #[cfg(test)]
